deepfake-patch-audit/datasets/frame_dataset.py
"""Frame-based dataset loader for video sequences."""

import os
from pathlib import Path
import numpy as np
from PIL import Image
import torch
from torch.utils.data import Dataset


class FrameDataset(Dataset):
    """
    Dataset for loading video frames from frame directories.
    Each directory contains sequential frames from a video.
    """

    def __init__(
        self,
        root_dir: str,
        split: str = "train",
        frames_per_video: int = 10,
        resize_size: int = 224,
        normalize: bool = True,
    ):
        """
        Args:
            root_dir: Root directory with 'real' and 'fake' subdirectories
            split: 'train', 'val', or 'test'
            frames_per_video: Number of frames to sample from each video
            resize_size: Target image size
            normalize: Whether to apply normalization
        """
        self.root_dir = Path(root_dir)
        self.split = split
        self.frames_per_video = frames_per_video
        self.resize_size = resize_size
        self.normalize = normalize

        self.samples = []
        self._load_samples()

    def _load_samples(self):
        """Load frame sequences."""
        # Implementation placeholder for loading frame sequences
        pass

    def __len__(self):
        return len(self.samples)

    def __getitem__(self, idx):
        """Load frame sequence and return stacked tensor."""
        # Implementation placeholder
        pass
